requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.74
botocore==1.42.74
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
dj-database-url==3.1.2
dj-email-url==1.0.6
dj-rest-auth==7.1.1
Django==6.0.2
django-allauth==65.14.3
django-cache-url==3.4.6
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
environs==14.6.0
fsspec==2026.2.0
gotrue==1.3.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
marshmallow==4.2.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
postgrest==0.11.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==1.0.6
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
squareup==44.0.1.20260122
storage3==0.6.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==1.2.0
supabase-auth==2.28.3
supabase-functions==2.28.3
supafunc==0.2.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==12.0
whitenoise==6.12.0
yarl==1.23.0
zstandard==0.25.0
